Move primitive UTxO types to separate module.
